Do not auto-update Copyright year
because the script cannot know if anything copyright-worthy has been done.

Plus removing old years can be seen as a sign that
we do not want to claim copyright for the old code anymore
and we do not want to give that impression.
The recommendation there is to either not give a year at all
(because we have version control anyway)
or specify multiple years, e.g. 2010-2012 or 2010,2012,2025